+++ sandbox.h
@@ -305,4 +305,4 @@ int cli_conn(const char *sock_name);
 ssize_t	readn(int fd, void *vptr, size_t n);
 ssize_t writen(int fd, const void *vptr, size_t n);
 ssize_t read_sandbox_message_header(int fd, uint16_t *version,
-				    uint16_t *id, uint64_t *len);
+				    uint16_t *id, uint32_t *len);
